requirements.txt
pyodbc>=4.0.39
python-dotenv>=1.0.0
requests>=2.31.0
requests-negotiate-sspi>=0.5.2
pytest>=7.0.0
